Move new analysis tests
