+++ tests/analyses/parsers/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for the OpticStudio analyses."""
