+++ tests/analyses/test_base.py
@@ -1,135 +1,353 @@
+from __future__ import annotations
+
+import inspect
+import json
+from dataclasses import fields
 from datetime import datetime
+from types import SimpleNamespace
 
 import numpy as np
-import pandas as pd
 import pytest
+from pandas import DataFrame
+from pydantic import Field
+from pydantic.dataclasses import dataclass
+from pydantic.fields import FieldInfo
 
+from zospy import constants
 from zospy.analyses.base import (
-    AnalysisMessage,
+    AnalysisData,
     AnalysisMetadata,
     AnalysisResult,
-    AttrDict,
-    OnComplete,
-    _AnalysisResultJSONDecoder,
-    _AnalysisResultJSONEncoder,
-    new_analysis,
+    AnalysisSettings,
+    BaseAnalysisWrapper,
+    _validated_setter,
 )
-from zospy.api import constants
+from zospy.analyses.decorators import analysis_settings
+from zospy.analyses.parsers.types import ValidatedDataFrame
+from zospy.analyses.reports.surface_data import SurfaceDataSettings
+from zospy.analyses.systemviewers.base import SystemViewerWrapper
 
 
-def test_oncomplete():
-    assert OnComplete.Close == "Close"
-    assert OnComplete.Sustain == "Sustain"
-    assert OnComplete.Release == "Release"
-
-    with pytest.raises(ValueError, match="'NonExistentValue' is not a valid OnComplete"):
-        OnComplete("NonExistentValue")
+def all_subclasses(cls):
+    return set(cls.__subclasses__()).union([s for c in cls.__subclasses__() for s in all_subclasses(c)])
 
 
-class TestToJSON:
-    @pytest.fixture(scope="class")
-    def example_analysis_result(self):
-        metadata = AnalysisMetadata(
-            DateTime=datetime.now(),
-            LensFile="C:\\Path\\To\\Lens\\File.ZOS",
-            LensTitle="LensTitle",
-            FeatureDescription="",
+analysis_wrapper_classes = all_subclasses(BaseAnalysisWrapper)
+analysis_wrapper_classes.remove(SystemViewerWrapper)
+
+
+class TestValidatedSetter:
+    class MockSettings:
+        int_setting: int = 1
+        string_setting: str = "a"
+
+    def test_get_existing(self):
+        settings = _validated_setter(self.MockSettings())
+
+        assert settings.int_setting == self.MockSettings.int_setting
+
+    def test_get_non_existing(self):
+        settings = _validated_setter(self.MockSettings())
+
+        with pytest.raises(AttributeError):
+            assert settings.non_existing
+
+    def test_set_existing(self):
+        settings = _validated_setter(self.MockSettings())
+
+        settings.int_setting = 2
+
+        assert settings.int_setting == 2
+
+    def test_set_non_existing(self):
+        settings = _validated_setter(self.MockSettings())
+
+        with pytest.raises(AttributeError, match="'MockSettings' object has no attribute 'non_existing'"):
+            settings.non_existing = 2
+
+
+@dataclass
+class MockAnalysisData:
+    int_data: int = 1
+    string_data: str = "a"
+
+
+@analysis_settings
+class MockAnalysisSettings:
+    int_setting: int = Field(default=1, description="An integer setting")
+    string_setting: str = Field(default="a", description="A string setting")
+
+
+class MockAnalysis(BaseAnalysisWrapper[MockAnalysisData, MockAnalysisSettings]):
+    TYPE = "MockAnalysis"
+
+    def __init__(
+        self,
+        *,
+        int_setting: int = 1,
+        string_setting: str = "a",
+        block_remove_temp_files: bool = False,
+    ):
+        super().__init__(settings_kws=locals())
+
+        self.block_remove_temp_files = block_remove_temp_files
+
+    def _create_analysis(self):
+        self._analysis = SimpleNamespace(
+            metadata=AnalysisMetadata(DateTime=datetime.now(), LensFile="", LensTitle="", FeatureDescription=""),
+            header_data=None,
+            messages=[],
+            Close=lambda: None,
         )
-        messages = [AnalysisMessage(ErrorCode="EC", Message="message")]
 
-        return AnalysisResult(
-            analysistype="DummyAnalysis",
-            data=AttrDict(
-                a=pd.Series({"b": 1, "c": 2, "d": 3, "e": 4}),
-                f=pd.DataFrame({"g": [5.6, 7.8, 9.10, 11.12], "h": ["i", "j", "k", "l"]}),
-                m="a useless string",
-            ),
-            metadata=metadata,
-            messages=messages,
+    def run_analysis(self) -> AnalysisData:
+        if self.block_remove_temp_files:
+            self._remove_config_file = False
+            self._remove_text_output_file = False
+
+        return MockAnalysisData()
+
+
+MockOpticStudioSystem = type("OpticStudioSystem", (), {})
+
+
+class TestAnalysisWrapper:
+    @staticmethod
+    def get_settings_defaults(settings_class):
+        result = {}
+
+        for field in fields(settings_class):
+            if isinstance(field.default, FieldInfo):
+                result[field.name] = field.default.default
+            else:
+                result[field.name] = field.default
+
+        return result
+
+    def test_get_settings_type(self):
+        assert MockAnalysis._settings_type == MockAnalysisSettings  # noqa: SLF001
+
+    def test_settings_type_is_specified(self):
+        assert MockAnalysis._settings_type is not AnalysisSettings  # noqa: SLF001
+
+    @pytest.mark.parametrize("cls", analysis_wrapper_classes)
+    def test_analyses_correct_analysis_name(self, cls):
+        assert cls.TYPE is not None
+        assert hasattr(constants.Analysis.AnalysisIDM, cls.TYPE)
+
+    @pytest.mark.parametrize("cls", analysis_wrapper_classes)
+    def test_init_all_keyword_only_parameters(self, cls):
+        assert all(p.kind.name == "KEYWORD_ONLY" for _, p in inspect.signature(cls).parameters.items())
+
+    @pytest.mark.parametrize("cls", analysis_wrapper_classes)
+    def test_init_contains_all_settings(self, cls):
+        if cls().settings is None:
+            return
+
+        init_signature = inspect.signature(cls.__init__)
+        settings_fields = fields(cls().settings)
+
+        assert all(field.name in init_signature.parameters for field in settings_fields)
+
+    @pytest.mark.parametrize("cls", analysis_wrapper_classes)
+    def test_analyses_default_values(self, cls):
+        if cls().settings is None:
+            return
+
+        settings_defaults = self.get_settings_defaults(type(cls().settings))
+        init_signature = inspect.signature(cls.__init__)
+
+        for field_name, default_value in settings_defaults.items():
+            assert field_name in init_signature.parameters
+            assert init_signature.parameters[field_name].default == default_value
+
+    def test_change_settings_from_parameters(self):
+        analysis = MockAnalysis(int_setting=2, string_setting="b")
+
+        assert analysis.settings.int_setting == 2
+        assert analysis.settings.string_setting == "b"
+
+    def test_change_settings_from_object(self):
+        settings = MockAnalysisSettings(int_setting=2, string_setting="b")
+        analysis = MockAnalysis.with_settings(settings)
+
+        assert analysis.settings.int_setting == 2
+        assert analysis.settings.string_setting == "b"
+
+    def test_settings_object_is_copied(self):
+        settings = MockAnalysisSettings(int_setting=2, string_setting="b")
+        analysis = MockAnalysis.with_settings(settings)
+
+        assert analysis.settings is not settings
+        assert analysis.settings == settings
+
+    def test_update_settings_object(self):
+        analysis = MockAnalysis(int_setting=1, string_setting="a")
+
+        analysis.update_settings(settings=MockAnalysisSettings(int_setting=2, string_setting="b"))
+
+        assert analysis.settings.int_setting == 2
+        assert analysis.settings.string_setting == "b"
+
+    def test_update_settings_dictionary(self):
+        analysis = MockAnalysis(int_setting=1, string_setting="a")
+
+        analysis.update_settings(settings_kws={"int_setting": 2, "string_setting": "b"})
+
+        assert analysis.settings.int_setting == 2
+        assert analysis.settings.string_setting == "b"
+
+    def test_update_settings_object_and_dictionary(self):
+        analysis = MockAnalysis(int_setting=1, string_setting="a")
+
+        analysis.update_settings(
+            settings=MockAnalysisSettings(int_setting=2, string_setting="a"), settings_kws={"string_setting": "b"}
         )
 
-    def test_restore_metadata(self, example_analysis_result):
-        restored_result = AnalysisResult.from_json(example_analysis_result.to_json())
+        assert analysis.settings.int_setting == 2
+        assert analysis.settings.string_setting == "b"
 
-        assert isinstance(restored_result.MetaData, AnalysisMetadata)
-        assert example_analysis_result.MetaData == restored_result.MetaData
+    def test_update_settings_no_dataclass_raises_type_error(self):
+        with pytest.raises(TypeError, match="settings should be a dataclass"):
+            MockAnalysis().update_settings(settings=123)
 
-    def test_restore_messages(self, example_analysis_result):
-        restored_result = AnalysisResult.from_json(example_analysis_result.to_json())
+    @pytest.mark.parametrize(
+        "temp_file_type,filename",
+        [
+            ("text_output_file", "test.txt"),
+            ("text_output_file", None),
+            ("config_file", "test.CFG"),
+            ("config_file", None),
+        ],
+    )
+    def test_create_temp_file(self, temp_file_type, filename, tmp_path, monkeypatch):
+        analysis = MockAnalysis(block_remove_temp_files=True)
+        monkeypatch.setattr(analysis, f"_needs_{temp_file_type}", True)
 
-        assert isinstance(restored_result.Messages[0], AnalysisMessage)
-        assert example_analysis_result.Messages == restored_result.Messages
+        if filename:
+            path = tmp_path / filename
+            path.touch()
+        else:
+            path = None
 
-    def test_restore_attrdict(self, example_analysis_result):
-        restored_result = AnalysisResult.from_json(example_analysis_result.to_json())
+        analysis.run(oss=MockOpticStudioSystem(), **{temp_file_type: path})
 
-        assert isinstance(restored_result.Data, AttrDict)
+        if filename:
+            assert getattr(analysis, temp_file_type) == path
 
-    def test_restore_pandas_series(self):
-        series = pd.Series({"a": 1, "b": 2, "c": 3}, dtype=float)
+        assert getattr(analysis, temp_file_type).exists()
+        assert getattr(analysis, temp_file_type).is_file()
 
-        restored_series = _AnalysisResultJSONDecoder().decode(_AnalysisResultJSONEncoder().encode(series))
+    @pytest.mark.parametrize(
+        "temp_file_type,filename",
+        [
+            ("text_output_file", "test.ABC"),
+            ("text_output_file", "test.TXT"),
+            ("config_file", "test.ABC"),
+            ("config_file", "test.cfg"),
+        ],
+    )
+    def test_temp_file_wrong_filename(self, temp_file_type, tmp_path, filename, monkeypatch):
+        analysis = MockAnalysis()
+        monkeypatch.setattr(analysis, f"_needs_{temp_file_type}", True)
 
-        assert isinstance(restored_series, pd.Series)
-        assert all(restored_series == series)
+        with pytest.raises(ValueError, match="File path should end with ."):
+            analysis.run(oss=MockOpticStudioSystem(), **{temp_file_type: tmp_path / filename})
 
-    def test_restore_pandas_multi_index_series(self):
-        index = pd.MultiIndex.from_product((["a", "b"], [1, 2]))
-        series = pd.Series([123, 456, 345, 678], index=index)
+    @pytest.mark.parametrize(
+        "temp_file_type,filename",
+        [
+            ("text_output_file", "test.txt"),
+            ("text_output_file", None),
+            ("config_file", "test.CFG"),
+            ("config_file", None),
+        ],
+    )
+    def test_remove_temp_file(self, temp_file_type, filename, tmp_path, monkeypatch):
+        analysis = MockAnalysis()
+        monkeypatch.setattr(analysis, f"_needs_{temp_file_type}", True)
 
-        restored_series = _AnalysisResultJSONDecoder().decode(_AnalysisResultJSONEncoder().encode(series))
+        if filename:
+            path = tmp_path / filename
+            path.touch()
+        else:
+            path = None
 
-        assert isinstance(restored_series, pd.Series)
-        assert isinstance(restored_series.index, pd.MultiIndex)
-        assert all(restored_series == series)
+        analysis.run(oss=MockOpticStudioSystem(), **{temp_file_type: path})
 
-    def test_restore_pandas_dataframe(self):
-        dataframe = pd.DataFrame(
-            {
-                "int_column": [1, 2, 3, 4],
-                "float_column": [1.2, 3.4, 5.6, 7.8],
-                "str_column": ["this", "ain't", "string", "theory"],
-            }
+        if path:
+            assert path.exists()
+        else:
+            assert not getattr(analysis, temp_file_type).exists()
+
+
+class TestAnalysisResultJSONConversion:
+    settings = SurfaceDataSettings()
+
+    # Only test for non-dataclass results, because dataclass results are tested separately in the corresponding
+    # analysis' tests.
+    @pytest.mark.parametrize(
+        "result_type,result_value,type_info",
+        [
+            (DataFrame, DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]}), {"data_type": "dataframe"}),
+            (np.ndarray, np.array([1, 2, 3]), {"data_type": "ndarray"}),
+        ],
+    )
+    def test_data_to_json(self, result_type, result_value, type_info):
+        result = AnalysisResult[result_type, MockAnalysisSettings](
+            data=result_value,
+            settings=MockAnalysisSettings(),
+            metadata=AnalysisMetadata(datetime.now(), "", "", ""),
+            header=None,
+            messages=None,
         )
 
-        restored_dataframe = _AnalysisResultJSONDecoder().decode(_AnalysisResultJSONEncoder().encode(dataframe))
+        result_json = result.to_json()
+        result_dict = json.loads(result_json)
 
-        assert isinstance(restored_dataframe, pd.DataFrame)
-        assert all(restored_dataframe == dataframe)
+        assert "__analysis_data__" in result_dict
+        assert result_dict["__analysis_data__"] == type_info
 
-    def test_restore_numpy_ndarray(self):
-        array = np.array([1.2345, 2.3456, 3.4567, 4.5678])
+    def test_settings_to_json(self):
+        result = AnalysisResult[ValidatedDataFrame, MockAnalysisSettings](
+            data=DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]}),
+            settings=MockAnalysisSettings(),
+            metadata=AnalysisMetadata(datetime.now(), "", "", ""),
+            header=None,
+            messages=None,
+        )
 
-        restored_array = _AnalysisResultJSONDecoder().decode(_AnalysisResultJSONEncoder().encode(array))
+        result_json = result.to_json()
+        result_dict = json.loads(result_json)
 
-        assert isinstance(restored_array, np.ndarray)
-        assert all(restored_array == array)
+        assert "__analysis_settings__" in result_dict
+        assert result_dict["__analysis_settings__"] == {
+            "data_type": "dataclass",
+            "name": "MockAnalysisSettings",
+            "module": "tests.analyses.test_base",
+        }
 
-    def test_restore_datetime(self):
-        test_datetime = datetime(year=2023, month=4, day=17, hour=10, minute=41, second=42, microsecond=43)
+    @pytest.mark.parametrize(
+        "result_type,result_value",
+        [
+            (ValidatedDataFrame, DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]})),
+            (np.ndarray, np.array([1, 2, 3])),
+        ],
+    )
+    def test_roundtrip(self, result_type, result_value):
+        result = AnalysisResult[result_type, MockAnalysisSettings](
+            data=result_value,
+            settings=self.settings,
+            metadata=AnalysisMetadata(datetime.now(), "", "", ""),
+            header=None,
+            messages=None,
+        )
 
-        restored_datetime = _AnalysisResultJSONDecoder().decode(_AnalysisResultJSONEncoder().encode(test_datetime))
+        result_json = result.to_json()
 
-        assert isinstance(restored_datetime, datetime)
-        assert restored_datetime == test_datetime
-
-
-def test_analysis_wrapper_dir(oss):
-    opticstudio_analysis_attributes = [
-        "GetSettings",
-        "GetResults",
-        "IsRunning",
-        "Apply",
-        "ApplyAndWaitForCompletion",
-        "Terminate",
-        "WaitForCompletion",
-        "Close",
-        "Release",
-        "ToFile",
-    ]
-
-    analysis = new_analysis(oss, analysis_type=constants.Analysis.AnalysisIDM.RayFan)
-
-    # Assert for each attribute separately, but in a single test so `oss` does not need to be recreated for every case.
-    for attribute in opticstudio_analysis_attributes:
-        assert attribute in dir(analysis)
+        result_roundtrip = AnalysisResult.from_json(result_json)
+        assert all(result_roundtrip.data == result.data)
+        assert result_roundtrip.settings == result.settings
+        assert result_roundtrip.metadata == result.metadata
+        assert result_roundtrip.header == result.header
+        assert result_roundtrip.messages == result.messages
