+++ zospy/analyses/parsers/base.py
@@ -35,7 +35,7 @@ def load_grammar(name: str) -> Lark:
     """
     try:
         return Lark.open_from_package(
-            "zospy.analyses.new.parsers.grammars", f"{name}.lark", parser="earley", start="start"
+            "zospy.analyses.parsers.grammars", f"{name}.lark", parser="earley", start="start"
         )
     except (FileNotFoundError, OSError) as e:
         raise FileNotFoundError(f"Grammar file {name}.lark not found") from e
